refactor out codes_nearest
