+++ src/intermediate_bindings/grib_nearest.rs
@@ -5,9 +5,7 @@ use eccodes_sys::{codes_handle, codes_nearest, CODES_NEAREST_SAME_DATA, CODES_NE
 use num_traits::FromPrimitive;
 
 use crate::{
-    codes_handle::NearestGridpoint,
-    errors::{CodesError, CodesInternal},
-    pointer_guard,
+    errors::{CodesError, CodesInternal}, pointer_guard, NearestGridpoint
 };
 
 pub unsafe fn codes_grib_nearest_new(
