+++ src/lib.rs
@@ -220,17 +220,16 @@ pub mod codes_handle;
 #[cfg(feature = "experimental_index")]
 #[cfg_attr(docsrs, doc(cfg(feature = "experimental_index")))]
 pub mod codes_index;
+pub mod codes_nearest;
 pub mod errors;
 mod intermediate_bindings;
 mod pointer_guard;
 
-pub use codes_handle::{
-    CodesHandle, CodesNearest, Key, KeyType, KeyedMessage, KeysIteratorFlags, NearestGridpoint,
-    ProductKind,
-};
+pub use codes_handle::{CodesHandle, Key, KeyType, KeyedMessage, KeysIteratorFlags, ProductKind};
 #[cfg(feature = "experimental_index")]
 #[cfg_attr(docsrs, doc(cfg(feature = "experimental_index")))]
 pub use codes_index::CodesIndex;
+pub use codes_nearest::{CodesNearest, NearestGridpoint};
 pub use errors::CodesError;
 pub use fallible_iterator::{FallibleIterator, IntoFallibleIterator};
 pub use fallible_streaming_iterator::FallibleStreamingIterator;
